itest: refactor watchtower related tests
